{{ cookiecutter.repo_name }}/requirements/base.txt
https://www.djangoproject.com/download/1.7.b4/tarball/
Jinja2==2.7.2
MarkupSafe==0.23
Pygments==1.6
Sphinx==1.2.2
dj-database-url==0.3.0
django-braces==1.4.0
django-configurations==0.8
django-crispy-forms==1.4.0
django-grappelli==2.5.3
docutils==0.11
envdir==0.6.1
psycopg2==2.5.3
pytz==2014.3
six==1.6.1
wheel==0.23.0
